+++ api.js
@@ -1,7 +1,0 @@
-/**
- * storage-node API modules
- */
-
-let { etl } = require("./storage/node/node-api/etl");
-
-exports.etl = etl;